# perf: use mimalloc for bindings_napi