--- a/rust/bindings_napi/Cargo.toml
+++ b/rust/bindings_napi/Cargo.toml
@@ -15,5 +15,11 @@ napi-derive = "2.13.0"
 parse_ast = { path = "../parse_ast" }
 xxhash = { path = "../xxhash" }
 
+[target.'cfg(not(target_os = "linux"))'.dependencies]
+mimalloc-rust = { version = "0.2" }
+
+[target.'cfg(all(target_os = "linux", not(all(target_env = "musl", target_arch = "aarch64"))))'.dependencies]
+mimalloc-rust = { version = "0.2", features = ["local-dynamic-tls"] }
+
 [build-dependencies]
 napi-build = "2.0.1"